+++ Wiki/ReleaseNotes.md
@@ -2,6 +2,11 @@
 
 ## vNext
 
+## 7.7.6
+
+- Parametrized timeout for GuaranteedDeliveryBroadcastBlock
+- Changed interception in mongodb, breaking changes in the interface.
+
 ## 7.7.5
 
 - Fixed concurrent bug in AtomicReadmodelFactory, added fasterflect to create readmodels.